Rename PAX files to header to make intellisense work
Otherwise, all includes from the `include` folder were marked as erros
